ENT-14284 - Test refactoring and minor fixes
