+++ Tokens/stockpaydividend/bridging-contracts/src/main/kotlin/com/r3/corda/lib/tokens/bridging/contracts/BridgingContract.kt
@@ -37,7 +37,7 @@ class BridgingContract : Contract {
         require(moveCommands.size == 1) { "Bridging must have one move command to lock token" }
 
         val lockedSum = tx.outputsOfType<FungibleToken>()
-            .filter { it.holder == bridgingCommand.bridgeAuthority } // TODO this is mute point for now, change to != bridgeAuthority, to filter only states owned by CI ...
+            .filter { it.holder != bridgingCommand.bridgeAuthority }
             // ... currently can't distinguish between locked and a change, both are for same holder
             .sumOf {
                 it.amount.toDecimal().toLong()
