+++ Tokens/stockpaydividend/bridging-flows/src/main/kotlin/com/r3/corda/lib/tokens/bridging/flows/BridgingAuthorityBootstrapService.kt
@@ -2,16 +2,23 @@ package com.r3.corda.lib.tokens.bridging.flows
 
 import com.r3.corda.lib.tokens.contracts.states.FungibleToken
 import net.corda.core.contracts.StateAndRef
+import net.corda.core.identity.AbstractParty
 import net.corda.core.identity.PartyAndCertificate
 import net.corda.core.node.AppServiceHub
 import net.corda.core.node.services.CordaService
 import net.corda.core.serialization.SingletonSerializeAsToken
+import net.corda.core.utilities.debug
+import org.slf4j.LoggerFactory
 import java.util.*
+import java.util.concurrent.Executors
 
 @CordaService
 class BridgingAuthorityBootstrapService(appServiceHub: AppServiceHub) : SingletonSerializeAsToken() {
-    val holdingIdentityPartyAndCertificate: PartyAndCertificate
-    val bridgeAuthorityParty = appServiceHub.ourIdentity()
+    private val holdingIdentityPartyAndCertificate: PartyAndCertificate
+    private val bridgeAuthorityParty = appServiceHub.myInfo.legalIdentities.first()
+    private val logger = LoggerFactory.getLogger(BridgingAuthorityBootstrapService::class.java)
+
+    private val executor = Executors.newSingleThreadExecutor()
 
     init {
         val cfg = appServiceHub.getAppContext().config
@@ -36,12 +43,18 @@ class BridgingAuthorityBootstrapService(appServiceHub: AppServiceHub) : Singleto
             }
         }
 
+        appServiceHub.registerUnloadHandler { onStop() }
         onStartup(appServiceHub)
+    }
+
+    private fun onStop() {
+        executor.shutdown()
     }
 
     private fun onStartup(appServiceHub: AppServiceHub) {
         //Retrieve states from receiver
         val receivedStates = appServiceHub.vaultService.queryBy(FungibleToken::class.java).states
+
 
         callFlow(receivedStates, appServiceHub)
         addVaultListener(appServiceHub)
@@ -56,16 +69,19 @@ class BridgingAuthorityBootstrapService(appServiceHub: AppServiceHub) : Singleto
 
     private fun callFlow(fungibleTokens: Collection<StateAndRef<FungibleToken>>, appServiceHub: AppServiceHub) {
         fungibleTokens.forEach { token ->
-                appServiceHub.startFlow(
-                    BridgeFungibleTokenFlow(
-                        holdingIdentityPartyAndCertificate.party,
-                        emptyList(),
-                        token,
-                        bridgeAuthorityParty
+            if (bridgeAuthorityParty !in previousOwnersOf(appServiceHub, token)) {
+                logger.debug { "Starting flow to bridge ${token.state.data.amount} to Solana" }
+                executor.submit {
+                    appServiceHub.startFlow(
+                        BridgeFungibleTokenFlow(
+                            holdingIdentityPartyAndCertificate.party,
+                            emptyList(),
+                            token,
+                            bridgeAuthorityParty
+                        )
                     )
-                )
+                }
             }
+        }
     }
-
-    private fun AppServiceHub.ourIdentity() = myInfo.legalIdentities.first()
 }
